+++ extension.json
@@ -64,7 +64,7 @@
 	},
 	"Hooks": {
 		"AbuseFilterShouldFilterAction": {
-			"handler": "Main"
+			"handler": "AbuseFilter"
 		},
 		"APIQuerySiteInfoGeneralInfo": {
 			"handler": "Main"
@@ -76,28 +76,28 @@
 			"handler": "Main"
 		},
 		"CreateWikiDeletion": {
-			"handler": "Main"
+			"handler": "CreateWiki"
 		},
 		"CreateWikiRename": {
-			"handler": "Main"
+			"handler": "CreateWiki"
 		},
 		"CreateWikiStatePrivate": {
-			"handler": "Main"
+			"handler": "CreateWiki"
 		},
 		"CreateWikiTables": {
-			"handler": "Main"
+			"handler": "CreateWiki"
 		},
 		"GetLocalURL::Internal": {
 			"handler": "Main"
 		},
 		"GlobalUserPageWikis": {
-			"handler": "Main"
+			"handler": "GlobalUserPage"
 		},
 		"ImportDumpJobAfterImport": {
-			"handler": "Main"
+			"handler": "ImportDump"
 		},
 		"ImportDumpJobGetFile": {
-			"handler": "Main"
+			"handler": "ImportDump"
 		},
 		"MessageCacheFetchOverrides": {
 			"handler": "Main"
@@ -128,16 +128,42 @@
 		}
 	},
 	"HookHandlers": {
+		"AbuseFilter": {
+			"class": "Miraheze\\MirahezeMagic\\HookHandlers\\AbuseFilter"
+		},
+		"CreateWiki": {
+			"class": "Miraheze\\MirahezeMagic\\HookHandlers\\CreateWiki",
+			"factory": "Miraheze\\MirahezeMagic\\HookHandlers\\CreateWiki::factory",
+			"services": [
+				"MainConfig",
+				"ConnectionProvider",
+				"FormatterFactory",
+				"ManageWikiModuleFactory",
+				"RepoGroup"
+			]
+		},
+		"GlobalUserPage": {
+			"class": "Miraheze\\MirahezeMagic\\HookHandlers\\GlobalUserPage",
+			"services": [
+				"CreateWikiDataStore",
+				"ManageWikiDataStoreFactory"
+			]
+		},
+		"ImportDump": {
+			"class": "Miraheze\\MirahezeMagic\\HookHandlers\\ImportDump",
+			"services": [
+				"ConnectionProvider"
+			]
+		},
 		"Main": {
 			"class": "Miraheze\\MirahezeMagic\\HookHandlers\\Main",
 			"factory": "Miraheze\\MirahezeMagic\\HookHandlers\\Main::factory",
 			"services": [
 				"MainConfig",
 				"CommentStore",
-				"ConnectionProvider",
 				"HttpRequestFactory",
 				"LinkRenderer",
-				"ManageWikiModuleFactory"
+				"TitleFactory"
 			]
 		},
 		"RequestWiki": {
