+++ includes/HookHandlers/Main.php
@@ -10,10 +10,6 @@ use MediaWiki\CommentStore\CommentStore;
 use MediaWiki\Config\Config;
 use MediaWiki\Config\GlobalVarConfig;
 use MediaWiki\Config\ServiceOptions;
-use MediaWiki\Context\RequestContext;
-use MediaWiki\Extension\AbuseFilter\AbuseFilterServices;
-use MediaWiki\Extension\AbuseFilter\Hooks\AbuseFilterShouldFilterActionHook;
-use MediaWiki\Extension\AbuseFilter\Variables\VariableHolder;
 use MediaWiki\Extension\CentralAuth\User\CentralAuthUser;
 use MediaWiki\Hook\BlockIpCompleteHook;
 use MediaWiki\Hook\ContributionsToolLinksHook;
@@ -26,44 +22,22 @@ use MediaWiki\Html\Html;
 use MediaWiki\Http\HttpRequestFactory;
 use MediaWiki\Linker\LinkRenderer;
 use MediaWiki\MainConfigNames;
-use MediaWiki\MediaWikiServices;
 use MediaWiki\Permissions\Hook\TitleReadWhitelistHook;
 use MediaWiki\Permissions\Hook\UserGetRightsRemoveHook;
-use MediaWiki\Shell\Shell;
 use MediaWiki\SpecialPage\SpecialPage;
 use MediaWiki\Title\Title;
+use MediaWiki\Title\TitleFactory;
 use MediaWiki\User\User;
 use MediaWiki\WikiMap\WikiMap;
-use Memcached;
 use MessageCache;
-use Miraheze\CreateWiki\Hooks\CreateWikiDeletionHook;
-use Miraheze\CreateWiki\Hooks\CreateWikiRenameHook;
-use Miraheze\CreateWiki\Hooks\CreateWikiStatePrivateHook;
-use Miraheze\CreateWiki\Hooks\CreateWikiTablesHook;
-use Miraheze\CreateWiki\Maintenance\SetContainersAccess;
-use Miraheze\ImportDump\Hooks\ImportDumpJobAfterImportHook;
-use Miraheze\ImportDump\Hooks\ImportDumpJobGetFileHook;
-use Miraheze\ImportDump\ImportDumpRequestManager;
-use Miraheze\ManageWiki\Helpers\Factories\ModuleFactory;
-use Redis;
 use Skin;
-use Throwable;
 use Wikimedia\IPUtils;
-use Wikimedia\Rdbms\DBConnRef;
-use Wikimedia\Rdbms\IConnectionProvider;
 
 class Main implements
-	AbuseFilterShouldFilterActionHook,
 	APIQuerySiteInfoGeneralInfoHook,
 	BlockIpCompleteHook,
 	ContributionsToolLinksHook,
-	CreateWikiDeletionHook,
-	CreateWikiRenameHook,
-	CreateWikiStatePrivateHook,
-	CreateWikiTablesHook,
 	GetLocalURL__InternalHook,
-	ImportDumpJobAfterImportHook,
-	ImportDumpJobGetFileHook,
 	MessageCacheFetchOverridesHook,
 	MimeMagicInitHook,
 	RecentChange_saveHook,
@@ -74,75 +48,38 @@ class Main implements
 {
 
 	public function __construct(
-		private readonly ServiceOptions $options,
 		private readonly CommentStore $commentStore,
-		private readonly IConnectionProvider $connectionProvider,
 		private readonly HttpRequestFactory $httpRequestFactory,
 		private readonly LinkRenderer $linkRenderer,
-		private readonly ModuleFactory $moduleFactory
+		private readonly TitleFactory $titleFactory,
+		private readonly ServiceOptions $options
 	) {
 	}
 
 	public static function factory(
 		Config $mainConfig,
 		CommentStore $commentStore,
-		IConnectionProvider $connectionProvider,
 		HttpRequestFactory $httpRequestFactory,
 		LinkRenderer $linkRenderer,
-		ModuleFactory $moduleFactory
+		TitleFactory $titleFactory
 	): self {
 		return new self(
+			$commentStore,
+			$httpRequestFactory,
+			$linkRenderer,
+			$titleFactory,
 			new ServiceOptions(
 				[
-					'CreateWikiCacheDirectory',
-					'EchoSharedTrackingDB',
-					'ManageWikiSettings',
 					'MirahezeMagicAccessIdsMap',
-					'MirahezeMagicMemcachedServers',
 					'MirahezeReportsBlockAlertKeywords',
 					'MirahezeReportsWriteKey',
 					MainConfigNames::ArticlePath,
-					MainConfigNames::JobTypeConf,
 					MainConfigNames::LanguageCode,
-					MainConfigNames::LocalDatabases,
 					MainConfigNames::Script,
 				],
 				$mainConfig
-			),
-			$commentStore,
-			$connectionProvider,
-			$httpRequestFactory,
-			$linkRenderer,
-			$moduleFactory
+			)
 		);
-	}
-
-	/**
-	 * Avoid filtering automatic account creation
-	 *
-	 * @param VariableHolder $vars
-	 * @param Title $title @phan-unused-param
-	 * @param User $user @phan-unused-param
-	 * @param array &$skipReasons
-	 * @return bool|void
-	 */
-	public function onAbuseFilterShouldFilterAction(
-		VariableHolder $vars,
-		Title $title,
-		User $user,
-		array &$skipReasons
-	) {
-		if ( defined( 'MW_PHPUNIT_TEST' ) ) {
-			return;
-		}
-
-		$varManager = AbuseFilterServices::getVariablesManager();
-
-		$action = $varManager->getVar( $vars, 'action', 1 )->toString();
-		if ( $action === 'autocreateaccount' ) {
-			$skipReasons[] = 'Blocking automatic account creation is not allowed';
-			return false;
-		}
 	}
 
 	/**
@@ -153,328 +90,7 @@ class Main implements
 		$result['miraheze'] = true;
 	}
 
-	/**
-	 * @inheritDoc
-	 * @param DBConnRef $cwdb @phan-unused-param
-	 */
-	public function onCreateWikiDeletion( DBConnRef $cwdb, string $dbname ): void {
-		global $wmgSwiftPassword, $wgGlobalUsageDatabase;
-
-		$dbw = $this->connectionProvider->getPrimaryDatabase(
-			$this->options->get( 'EchoSharedTrackingDB' )
-		);
-
-		$dbw->newDeleteQueryBuilder()
-			->deleteFrom( 'echo_unread_wikis' )
-			->where( [ 'euw_wiki' => $dbname ] )
-			->caller( __METHOD__ )
-			->execute();
-
-		if ( $wgGlobalUsageDatabase ) {
-			$gudDb = $this->connectionProvider->getPrimaryDatabase(
-				$wgGlobalUsageDatabase
-			);
-
-			$gudDb->newDeleteQueryBuilder()
-				->deleteFrom( 'globalimagelinks' )
-				->where( [ 'gil_wiki' => $dbname ] )
-				->caller( __METHOD__ )
-				->execute();
-		}
-
-		foreach ( $this->options->get( MainConfigNames::LocalDatabases ) as $db ) {
-			$mwSettings = $this->moduleFactory->settings( $db );
-
-			foreach ( $this->options->get( 'ManageWikiSettings' ) as $var => $setConfig ) {
-				if (
-					$setConfig['type'] === 'database' &&
-					$mwSettings->list( $var ) === $dbname
-				) {
-					$mwSettings->remove( [ $var ], default: null );
-					$mwSettings->commit();
-				}
-			}
-		}
-
-		$limits = [ 'memory' => 0, 'filesize' => 0, 'time' => 0, 'walltime' => 0 ];
-
-		// Get a list of containers to delete for the wiki
-		$containers = explode( "\n",
-			trim( Shell::command(
-				'swift', 'list',
-				'--prefix', 'miraheze-' . $dbname . '-',
-				'-A', 'https://swift-lb.wikitide.net/auth/v1.0',
-				'-U', 'mw:media',
-				'-K', $wmgSwiftPassword
-			)->limits( $limits )
-				->disableSandbox()
-				->execute()->getStdout()
-			)
-		);
-
-		foreach ( $containers as $container ) {
-			// Just an extra precaution to ensure we don't select the wrong containers
-			if ( !str_contains( $container, $dbname . '-' ) ) {
-				continue;
-			}
-
-			// Delete the container
-			Shell::command(
-				'swift', 'delete',
-				$container,
-				'--object-threads', '1',
-				'--container-threads', '1',
-				'-A', 'https://swift-lb.wikitide.net/auth/v1.0',
-				'-U', 'mw:media',
-				'-K', $wmgSwiftPassword
-			)->limits( $limits )
-				->disableSandbox()
-				->execute();
-		}
-
-		$this->removeRedisKey( "*{$dbname}*" );
-		$this->removeMemcachedKey( ".*{$dbname}.*" );
-	}
-
-	/**
-	 * @inheritDoc
-	 * @param DBConnRef $cwdb @phan-unused-param
-	 */
-	public function onCreateWikiRename(
-		DBConnRef $cwdb,
-		string $oldDbName,
-		string $newDbName
-	): void {
-		global $wmgSwiftPassword, $wgGlobalUsageDatabase;
-
-		$dbw = $this->connectionProvider->getPrimaryDatabase(
-			$this->options->get( 'EchoSharedTrackingDB' )
-		);
-
-		$dbw->newUpdateQueryBuilder()
-			->update( 'echo_unread_wikis' )
-			->set( [ 'euw_wiki' => $newDbName ] )
-			->where( [ 'euw_wiki' => $oldDbName ] )
-			->caller( __METHOD__ )
-			->execute();
-
-		if ( $wgGlobalUsageDatabase ) {
-			$gudDb = $this->connectionProvider->getPrimaryDatabase(
-				$wgGlobalUsageDatabase
-			);
-
-			$gudDb->newUpdateQueryBuilder()
-				->update( 'globalimagelinks' )
-				->set( [ 'gil_wiki' => $newDbName ] )
-				->where( [ 'gil_wiki' => $oldDbName ] )
-				->caller( __METHOD__ )
-				->execute();
-		}
-
-		foreach ( $this->options->get( MainConfigNames::LocalDatabases ) as $db ) {
-			$mwSettings = $this->moduleFactory->settings( $db );
-
-			foreach ( $this->options->get( 'ManageWikiSettings' ) as $var => $setConfig ) {
-				if (
-					$setConfig['type'] === 'database' &&
-					$mwSettings->list( $var ) === $oldDbName
-				) {
-					$mwSettings->modify( [ $var => $newDbName ], default: null );
-					$mwSettings->commit();
-				}
-			}
-		}
-
-		$limits = [ 'memory' => 0, 'filesize' => 0, 'time' => 0, 'walltime' => 0 ];
-
-		// Get a list of containers to download, and later upload for the wiki
-		$containers = explode( "\n",
-			trim( Shell::command(
-				'swift', 'list',
-				'--prefix', 'miraheze-' . $oldDbName . '-',
-				'-A', 'https://swift-lb.wikitide.net/auth/v1.0',
-				'-U', 'mw:media',
-				'-K', $wmgSwiftPassword
-			)->limits( $limits )
-				->disableSandbox()
-				->execute()->getStdout()
-			)
-		);
-
-		foreach ( $containers as $container ) {
-			// Just an extra precaution to ensure we don't select the wrong containers
-			if ( !str_contains( $container, $oldDbName . '-' ) ) {
-				continue;
-			}
-
-			// Get a list of all files in the container to ensure everything is present in new container later.
-			$oldContainerList = Shell::command(
-				'swift', 'list',
-				$container,
-				'-A', 'https://swift-lb.wikitide.net/auth/v1.0',
-				'-U', 'mw:media',
-				'-K', $wmgSwiftPassword
-			)->limits( $limits )
-				->disableSandbox()
-				->execute()->getStdout();
-
-			// Download the container
-			Shell::command(
-				'swift', 'download',
-				$container,
-				'-D', wfTempDir() . '/' . $container,
-				'-A', 'https://swift-lb.wikitide.net/auth/v1.0',
-				'-U', 'mw:media',
-				'-K', $wmgSwiftPassword
-			)->limits( $limits )
-				->disableSandbox()
-				->execute();
-
-			$newContainer = str_replace( $oldDbName, $newDbName, $container );
-
-			// Upload to new container
-			// We have to use exec here, as Shell::command does not work for this
-			// phpcs:ignore MediaWiki.Usage.ForbiddenFunctions
-			exec( escapeshellcmd(
-				implode( ' ', [
-					'swift', 'upload',
-					$newContainer,
-					wfTempDir() . '/' . $container,
-					'--object-name', '""',
-					'-A', 'https://swift-lb.wikitide.net/auth/v1.0',
-					'-U', 'mw:media',
-					'-K', $wmgSwiftPassword
-				] )
-			) );
-
-			wfDebugLog( 'MirahezeMagic', "Container '$newContainer' created." );
-
-			$newContainerList = Shell::command(
-				'swift', 'list',
-				$newContainer,
-				'-A', 'https://swift-lb.wikitide.net/auth/v1.0',
-				'-U', 'mw:media',
-				'-K', $wmgSwiftPassword
-			)->limits( $limits )
-				->disableSandbox()
-				->execute()->getStdout();
-
-			if ( $newContainerList === $oldContainerList ) {
-				// Everything has been correctly copied over
-				// wipe files from the temp directory and delete old container
-
-				// Delete the container
-				Shell::command(
-					'swift', 'delete',
-					$container,
-					'-A', 'https://swift-lb.wikitide.net/auth/v1.0',
-					'-U', 'mw:media',
-					'-K', $wmgSwiftPassword
-				)->limits( $limits )
-					->disableSandbox()
-					->execute();
-
-				wfDebugLog( 'MirahezeMagic', "Container '$container' deleted." );
-
-				// Wipe from the temp directory
-				Shell::command( '/bin/rm', '-rf', wfTempDir() . '/' . $container )
-					->limits( $limits )
-					->disableSandbox()
-					->execute();
-			} else {
-				/**
-				 * We need to log this, as otherwise all files may not have been succesfully
-				 * moved to the new container, and they still exist locally. We should know that.
-				 */
-				wfDebugLog( 'MirahezeMagic', "The rename of wiki {$oldDbName} to {$newDbName} may not have been successful. Files still exist locally in {wfTempDir()} and the Swift containers for the old wiki still exist." );
-			}
-		}
-
-		Shell::makeScriptCommand(
-			SetContainersAccess::class,
-			[ '--wiki', $newDbName ]
-		)->limits( $limits )->execute();
-
-		$this->removeRedisKey( "*{$oldDbName}*" );
-		$this->removeMemcachedKey( ".*{$oldDbName}.*" );
-	}
-
-	public function onCreateWikiStatePrivate( string $dbname ): void {
-		$localRepo = MediaWikiServices::getInstance()->getRepoGroup()->getLocalRepo();
-		$sitemaps = $localRepo->getBackend()->getTopFileList( [
-			'dir' => $localRepo->getZonePath( 'public' ) . '/sitemaps',
-			'adviseStat' => false,
-		] );
-
-		foreach ( $sitemaps as $sitemap ) {
-			$status = $localRepo->getBackend()->quickDelete( [
-				'src' => $localRepo->getZonePath( 'public' ) . "/sitemaps/$sitemap",
-			] );
-
-			if ( !$status->isOK() ) {
-				$statusFormatter = MediaWikiServices::getInstance()->getFormatterFactory()
-					->getStatusFormatter( RequestContext::getMain() );
-
-				/**
-				 * We need to log this, as otherwise the sitemaps may
-				 * not be being deleted for private wikis. We should know that.
-				 */
-				$statusMessage = $statusFormatter->getWikiText( $status );
-				wfDebugLog( 'MirahezeMagic', "Sitemap \"$sitemap\" failed to delete for $dbname: $statusMessage" );
-			}
-		}
-
-		$localRepo->getBackend()->clean( [ 'dir' => $localRepo->getZonePath( 'public' ) . '/sitemaps' ] );
-	}
-
-	public function onCreateWikiTables( array &$tables ): void {
-		$tables['localnames'] = 'ln_wiki';
-		$tables['localuser'] = 'lu_wiki';
-	}
-
-	/**
-	 * @inheritDoc
-	 * @param ImportDumpRequestManager $importDumpRequestManager @phan-unused-param
-	 */
-	public function onImportDumpJobAfterImport( $filePath, $importDumpRequestManager ): void {
-		$limits = [ 'memory' => 0, 'filesize' => 0, 'time' => 0, 'walltime' => 0 ];
-		Shell::command( '/bin/rm', $filePath )
-			->limits( $limits )
-			->disableSandbox()
-			->execute();
-	}
-
-	public function onImportDumpJobGetFile( &$filePath, $importDumpRequestManager ): void {
-		global $wmgSwiftPassword;
-
-		$dbr = $this->connectionProvider->getReplicaDatabase( 'virtual-importdump' );
-
-		$container = $dbr->getDomainID() === 'metawikibeta' ?
-			'miraheze-metawikibeta-local-public' :
-			'miraheze-metawiki-local-public';
-
-		$limits = [ 'memory' => 0, 'filesize' => 0, 'time' => 0, 'walltime' => 0 ];
-
-		Shell::command(
-			'swift', 'download',
-			$container,
-			$importDumpRequestManager->getSplitFilePath(),
-			'-o', $filePath,
-			'-A', 'https://swift-lb.wikitide.net/auth/v1.0',
-			'-U', 'mw:media',
-			'-K', $wmgSwiftPassword
-		)->limits( $limits )
-			->disableSandbox()
-			->execute();
-	}
-
-	/**
-	 * From WikimediaMessages
-	 * When core requests certain messages, change the key to a Miraheze version.
-	 *
-	 * @see https://www.mediawiki.org/wiki/Manual:Hooks/MessageCacheFetchOverrides
-	 * @param string[] &$keys
-	 */
+	/** @inheritDoc */
 	public function onMessageCacheFetchOverrides( array &$keys ): void {
 		static $keysToOverride = [
 			'centralauth-groupname',
@@ -542,11 +158,10 @@ class Main implements
 			'uploadtext',
 			'webauthn-module-description',
 			'wikibase-sitelinks-miraheze',
-			'vector-night-mode-issue-reporting-notice-url'
+			'vector-night-mode-issue-reporting-notice-url',
 		];
 
 		$languageCode = $this->options->get( MainConfigNames::LanguageCode );
-
 		$transformationCallback = static function ( string $key, MessageCache $cache ) use ( $languageCode ): string {
 			$transformedKey = "miraheze-$key";
 
@@ -579,8 +194,9 @@ class Main implements
 		}
 	}
 
+	/** @inheritDoc */
 	public function onTitleReadWhitelist( $title, $user, &$whitelisted ) {
-		if ( $title->equals( Title::newMainPage() ) ) {
+		if ( $title->equals( $this->titleFactory->newMainPage() ) ) {
 			$whitelisted = true;
 			return;
 		}
@@ -592,7 +208,7 @@ class Main implements
 			'CreateAccount',
 			'Notifications',
 			'OAuth',
-			'ResetPassword'
+			'ResetPassword',
 		];
 
 		if ( $user->isAllowed( 'interwiki' ) ) {
@@ -601,8 +217,7 @@ class Main implements
 
 		if ( $title->isSpecialPage() ) {
 			$rootName = strtok( $title->getText(), '/' );
-			$rootTitle = Title::makeTitle( $title->getNamespace(), $rootName );
-
+			$rootTitle = $this->titleFactory->makeTitle( $title->getNamespace(), $rootName );
 			foreach ( $specialsArray as $page ) {
 				if ( $rootTitle->equals( SpecialPage::getTitleFor( $page ) ) ) {
 					$whitelisted = true;
@@ -612,29 +227,12 @@ class Main implements
 		}
 	}
 
-	public function onGlobalUserPageWikis( array &$list ): bool {
-		$cwCacheDir = $this->options->get( 'CreateWikiCacheDirectory' );
-
-		if ( file_exists( "$cwCacheDir/databases.php" ) ) {
-			$databasesArray = include "$cwCacheDir/databases.php";
-			$dbList = array_keys( $databasesArray['databases'] ?? [] );
-
-			// Filter out those databases that don't have GlobalUserPage enabled
-			$list = array_filter( $dbList, function ( $dbname ) {
-				$mwExtensions = $this->moduleFactory->extensions( $dbname );
-				return in_array( 'globaluserpage', $mwExtensions->list(), true );
-			} );
-
-			return false;
-		}
-
-		return true;
-	}
-
+	/** @inheritDoc */
 	public function onMimeMagicInit( $mimeMagic ) {
 		$mimeMagic->addExtraTypes( 'text/plain txt off' );
 	}
 
+	/** @inheritDoc */
 	public function onSkinAddFooterLinks( Skin $skin, string $key, array &$footerItems ) {
 		if ( $key === 'places' ) {
 			$footerItems['termsofservice'] = $this->addFooterLink( $skin, 'termsofservice', 'termsofservicepage' );
@@ -642,6 +240,7 @@ class Main implements
 		}
 	}
 
+	/** @inheritDoc */
 	public function onUserGetRightsRemove( $user, &$rights ) {
 		// Remove read from global groups on some wikis
 		foreach ( $this->options->get( 'MirahezeMagicAccessIdsMap' ) as $wiki => $ids ) {
@@ -650,18 +249,18 @@ class Main implements
 
 				if ( $centralAuthUser &&
 					$centralAuthUser->exists() &&
-					!in_array( $centralAuthUser->getId(), $ids )
+					!in_array( $centralAuthUser->getId(), $ids, true )
 				) {
 					$rights = array_unique( $rights );
-					unset( $rights[array_search( 'read', $rights )] );
+					unset( $rights[array_search( 'read', $rights, true )] );
 				}
 			}
 		}
 	}
 
+	/** @inheritDoc */
 	public function onSiteNoticeAfter( &$siteNotice, $skin ) {
 		$cwConfig = new GlobalVarConfig( 'cw' );
-
 		if ( $cwConfig->get( 'Closed' ) ) {
 			if ( $cwConfig->get( 'Private' ) ) {
 				$siteNotice .= '<div class="wikitable" style="text-align: center; width: 90%; margin-left: auto; margin-right:auto; padding: 15px; border: 4px solid black; background-color: #EEE;"> <span class="plainlinks"> <img src="https://static.wikitide.net/metawiki/0/02/Wiki_lock.png" align="left" style="width:80px;height:90px;">' . $skin->msg( 'miraheze-sitenotice-closed-private' )->parse() . '</span></div>';
@@ -680,6 +279,7 @@ class Main implements
 	}
 
 	/**
+	 * @inheritDoc
 	 * phpcs:disable MediaWiki.NamingConventions.LowerCamelFunctionsName.FunctionName
 	 */
 	public function onRecentChange_save( $recentChange ) {
@@ -690,7 +290,7 @@ class Main implements
 		}
 
 		$globalUserGroups = CentralAuthUser::getInstanceByName( $recentChange->mAttribs['rc_user_text'] )->getGlobalGroups();
-		if ( !in_array( 'trustandsafety', $globalUserGroups ) ) {
+		if ( !in_array( 'trustandsafety', $globalUserGroups, true ) ) {
 			return;
 		}
 
@@ -712,10 +312,9 @@ class Main implements
 	public function onBlockIpComplete( $block, $user, $priorBlock ) {
 		// TODO: do we want to add localization support for these keywords, so they match in other languages as well?
 		$blockAlertKeywords = $this->options->get( 'MirahezeReportsBlockAlertKeywords' );
-
 		foreach ( $blockAlertKeywords as $keyword ) {
-			// use strtolower for case insensitivity
-			if ( str_contains( strtolower( $block->getReasonComment()->text ), strtolower( $keyword ) ) ) {
+			// use mb_strtolower for case insensitivity
+			if ( str_contains( mb_strtolower( $block->getReasonComment()->text ), mb_strtolower( $keyword ) ) ) {
 				$data = [
 					'writekey' => $this->options->get( 'MirahezeReportsWriteKey' ),
 					'username' => $block->getTargetName(),
@@ -725,7 +324,6 @@ class Main implements
 				];
 
 				$this->httpRequestFactory->post( 'https://reports.miraheze.org/api/report', [ 'postData' => $data ], __METHOD__ );
-
 				break;
 			}
 		}
@@ -737,32 +335,28 @@ class Main implements
 	 */
 	public function onContributionsToolLinks( $id, Title $title, array &$tools, SpecialPage $specialPage ) {
 		$username = $title->getText();
-
 		if ( !IPUtils::isIPAddress( $username ) ) {
 			$globalUserGroups = CentralAuthUser::getInstanceByName( $username )->getGlobalGroups();
 
 			if (
-				!in_array( 'steward', $globalUserGroups ) &&
-				!in_array( 'global-sysop', $globalUserGroups ) &&
+				!in_array( 'steward', $globalUserGroups, true ) &&
+				!in_array( 'global-sysop', $globalUserGroups, true ) &&
 				!$specialPage->getUser()->isAllowed( 'centralauth-lock' )
 			) {
 				return;
 			}
 
 			$tools['centralauth'] = $this->linkRenderer->makeExternalLink(
-				'https://meta.miraheze.org/wiki/Special:CentralAuth/' . $username,
-				strtolower( $specialPage->msg( 'centralauth' )->text() ),
+				"https://meta.miraheze.org/wiki/Special:CentralAuth/$username",
+				mb_strtolower( $specialPage->msg( 'centralauth' )->text() ),
 				SpecialPage::getTitleFor( 'CentralAuth' )
 			);
 		}
 	}
 
 	/**
+	 * @inheritDoc
 	 * phpcs:disable MediaWiki.NamingConventions.LowerCamelFunctionsName.FunctionName
-	 *
-	 * @param Title $title
-	 * @param string &$url
-	 * @param string $query
 	 */
 	public function onGetLocalURL__Internal( $title, &$url, $query ) {
 		// phpcs:enable
@@ -781,14 +375,16 @@ class Main implements
 		}
 	}
 
-	private function addFooterLink( $skin, $desc, $page ) {
+	private function addFooterLink( Skin $skin, string $desc, string $page ): string {
 		if ( $skin->msg( $desc )->inContentLanguage()->isDisabled() ) {
-			$title = null;
-		} else {
-			$title = Title::newFromText( $skin->msg( $page )->inContentLanguage()->text() );
+			return '';
 		}
 
-		if ( !$title ) {
+		$title = $this->titleFactory->newFromText(
+			$skin->msg( $page )->inContentLanguage()->text()
+		);
+
+		if ( $title === null ) {
 			return '';
 		}
 
@@ -796,56 +392,5 @@ class Main implements
 			[ 'href' => $title->fixSpecialName()->getLinkURL() ],
 			$skin->msg( $desc )->text()
 		);
-	}
-
-	/** Removes redis keys for jobrunner */
-	private function removeRedisKey( string $key ) {
-		$jobTypeConf = $this->options->get( MainConfigNames::JobTypeConf );
-		if ( !isset( $jobTypeConf['default']['redisServer'] ) || !$jobTypeConf['default']['redisServer'] ) {
-			return;
-		}
-
-		$hostAndPort = IPUtils::splitHostAndPort( $jobTypeConf['default']['redisServer'] );
-
-		if ( $hostAndPort ) {
-			try {
-				$redis = new Redis();
-				$redis->connect( $hostAndPort[0], $hostAndPort[1] );
-				$redis->auth( $jobTypeConf['default']['redisConfig']['password'] );
-				$redis->del( $redis->keys( $key ) );
-			} catch ( Throwable ) {
-				// empty
-			}
-		}
-	}
-
-	/** Remove memcached keys */
-	private function removeMemcachedKey( string $key ) {
-		$memcachedServers = $this->options->get( 'MirahezeMagicMemcachedServers' );
-
-		try {
-			foreach ( $memcachedServers as $memcachedServer ) {
-				$memcached = new Memcached();
-
-				$memcached->addServer( $memcachedServer[0], (string)$memcachedServer[1] );
-
-				// Fetch all keys
-				$keys = $memcached->getAllKeys();
-				if ( !is_array( $keys ) ) {
-					return;
-				}
-
-				foreach ( $keys as $item ) {
-					// Decide which keys to delete
-					if ( preg_match( "/{$key}/", $item ) ) {
-						$memcached->delete( $item );
-					} else {
-						continue;
-					}
-				}
-			}
-		} catch ( Throwable ) {
-			// empty
-		}
 	}
 }
