+++ includes/MirahezeIRCRCFeedFormatter.php
@@ -3,9 +3,8 @@
 namespace Miraheze\MirahezeMagic;
 
 use MediaWiki\Extension\CentralAuth\User\CentralAuthUser;
-use MediaWiki\MainConfigNames;
-use MediaWiki\MediaWikiServices;
 use MediaWiki\RCFeed\IRCColourfulRCFeedFormatter;
+use MediaWiki\WikiMap\WikiMap;
 use RecentChange;
 
 class MirahezeIRCRCFeedFormatter extends IRCColourfulRCFeedFormatter {
@@ -17,9 +16,7 @@ class MirahezeIRCRCFeedFormatter extends IRCColourfulRCFeedFormatter {
 			return null;
 		}
 
-		$mainConfig = MediaWikiServices::getInstance()->getMainConfig();
-		$dbname = $mainConfig->get( MainConfigNames::DBname );
-
+		$wikiId = WikiMap::getCurrentWikiId();
 		$attribs = $rc->getAttributes();
 
 		/**
@@ -27,17 +24,17 @@ class MirahezeIRCRCFeedFormatter extends IRCColourfulRCFeedFormatter {
 		 * WikiTide Foundation Trust and Safety team.
 		 */
 		if (
-			$attribs['rc_type'] == RC_LOG &&
+			$attribs['rc_type'] === RC_LOG &&
 			$attribs['rc_log_type'] === 'renameuser'
 		) {
 			$globalUserGroups = CentralAuthUser::getInstanceByName( $attribs['rc_user_text'] )->getGlobalGroups();
-			if ( in_array( 'trustandsafety', $globalUserGroups ) ) {
+			if ( in_array( 'trustandsafety', $globalUserGroups, true ) ) {
 				return null;
 			}
 		}
 
 		// Prefix is \003, no colour (\003) switches
 		// back to the term default.
-		return "$dbname \0035*\003 $lineFromParent";
+		return "$wikiId \0035*\003 $lineFromParent";
 	}
 }
